File structure refactor and work on UI components
