+++ app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import "./ui/globals.css";
+import "../styles/globals.css";
 
 export const metadata: Metadata = {
   title: {
