+++ app/page.tsx
@@ -1,8 +1,12 @@
-import UnderMaint from "./ui/under_maint/under_maint";
+import UnderMaint from "./ui/under_maint";
+
+// DEV
+import PunchClock from "./ui/punch_clock";
 
 export default function Home() {
   return (
     <div>
+      <PunchClock />
       <UnderMaint />
     </div>
   );
