+++ styles/globals.css
@@ -3,19 +3,19 @@
 @tailwind utilities;
 
 :root {
-  --background: #444560;
+  --datasmith: #444560;
   --foreground: #171717;
 }
 
 @media (prefers-color-scheme: dark) {
   :root {
-    --background: #444560;
+    --datasmith: #444560;
     --foreground: #ededed;
   }
 }
 
 body {
-  color: var(--foreground);
-  background: var(--background);
-  font-family: Arial, Helvetica, sans-serif;
+  color: var(--datasmith);
+  background: var(--foreground);
+  font-family: poppins;
 }
